+++ cmd/sbs-data/main.go
@@ -26,12 +26,12 @@ import (
 	"google.golang.org/grpc"
 )
 
-var buildVersion = namrbdversion.Current
+var buildVersion = namrbdversion.ProductVersion()
 
 func main() {
 	for _, arg := range os.Args[1:] {
 		if arg == "--version" || arg == "version" {
-			fmt.Println(buildVersion)
+			fmt.Println(namrbdversion.BuildSummary())
 			return
 		}
 	}